Refactor code :hammer:
