+++ src/api/playgroundMetrics.ts
@@ -1,5 +1,5 @@
 import { ApiClient } from 'middleware';
 
-export const nehaCallMetrics = ApiClient({
+export const getPlaygroundCallMetrics = ApiClient({
     url: `/v1/voice-chat/call/search`
 });
