+++ src/components/common/paginatedTable/index.ts
@@ -3,10 +3,6 @@ export { DisabledRowMask } from './DisabledRowMask';
 export { PaginatedTableHeader } from './PaginatedTableHeader';
 export { PaginatedTableSkeleton } from './PaginatedTableSkeleton';
 export { TextOverFlow } from './TextOverflow';
-export { Transcript } from './Transcript';
-export { Result } from './Result';
-export { Context } from './Context';
-export { CallStatus } from './CallStatus';
 export {
     dynamicInputCellSx,
     getDisabledRowMaskSx
