--- src/components/neha/index.ts
@@ -1,1 +1,0 @@
-export { CallMetrics } from './CallMetrics';
