+++ src/components/playgroundMetrics/index.ts
@@ -1,0 +1,2 @@
+export * from './tableCells';
+export { VoiceCallMetrics } from './VoiceCallMetrics';
