+++ src/containers/PlaygroundMetricsContainer.tsx
@@ -1,12 +1,12 @@
 import { Box, Center } from '@chakra-ui/react';
-import { CallMetrics } from '@components/neha';
+import { VoiceCallMetrics } from '@components/playgroundMetrics';
 import { NAVBAR_HEIGHT } from 'Constants';
 
-export const NehaMetricsContainer = () => {
+export const PlaygroundMetricsContainer = () => {
     return (
         <Box height={`calc(100vh - ${NAVBAR_HEIGHT})`} overflow="scroll">
             <Center flexDirection="column" py={4} px={8}>
-                <CallMetrics />
+                <VoiceCallMetrics />
             </Center>
         </Box>
     );
